backend/tests/test_auth.py
"""
后端认证模块测试

测试登录注册相关API接口
"""
import pytest
from unittest.mock import AsyncMock, MagicMock, patch
from datetime import datetime

from fastapi.testclient import TestClient
from httpx import AsyncClient, ASGITransport

import sys
import os
sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))

from app.main import app
from app.models.user import User, StudentProfile
from app.schemas.auth import TokenResponse


# 测试客户端
@pytest.fixture
def client():
    """创建测试客户端"""
    return TestClient(app)


# Mock数据库会话
@pytest.fixture
def mock_db():
    """Mock数据库会话"""
    db = AsyncMock()
    db.execute = AsyncMock()
    db.commit = AsyncMock()
    db.rollback = AsyncMock()
    db.close = AsyncMock()
    return db


# Mock用户数据
@pytest.fixture
def mock_user():
    """Mock用户数据"""
    user = MagicMock(spec=User)
    user.id = 1
    user.openid = "test_openid_123"
    user.unionid = None
    user.nickname = "测试用户"
    user.phone = "13800138000"
    user.avatar_url = "https://example.com/avatar.jpg"
    user.is_active = 1
    user.created_at = datetime.now()
    return user


# Mock Token数据
@pytest.fixture
def mock_token_response():
    """Mock Token响应"""
    return TokenResponse(
        access_token="mock_jwt_token_here",
        token_type="bearer",
        expires_in=10080
    )


class TestAuthAPI:
    """认证API测试类"""

    def test_health_check(self, client):
        """测试健康检查端点"""
        response = client.get("/health")
        assert response.status_code == 200
        assert response.json() == {"status": "healthy"}

    def test_root_endpoint(self, client):
        """测试根路径端点"""
        response = client.get("/")
        assert response.status_code == 200
        data = response.json()
        assert data["message"] == "AI小商 API"
        assert data["version"] == "1.0.0"


class TestLoginEndpoint:
    """登录端点测试"""

    def test_login_returns_501(self, client):
        """测试登录接口返回501 - 待实现"""
        response = client.post("/api/auth/login", json={"code": "test_code"})
        assert response.status_code == 501
        assert response.json()["detail"] == "微信登录待实现"

    def test_login_invalid_payload(self, client):
        """测试登录接口无效载荷"""
        response = client.post("/api/auth/login", json={})
        # Pydantic验证失败返回422
        assert response.status_code == 422


class TestRegisterEndpoint:
    """注册端点测试"""

    def test_register_returns_501(self, client):
        """测试注册接口返回501 - 待实现"""
        response = client.post(
            "/api/auth/register",
            json={
                "openid": "test_openid",
                "nickname": "测试用户"
            }
        )
        assert response.status_code == 501
        assert response.json()["detail"] == "用户注册待实现"

    def test_register_missing_fields(self, client):
        """测试注册接口缺少必填字段"""
        response = client.post("/api/auth/register", json={"nickname": "测试"})
        # Pydantic验证失败返回422
        assert response.status_code == 422


class TestGetCurrentUser:
    """获取当前用户测试"""

    def test_get_current_user_returns_501(self, client):
        """测试获取当前用户返回501 - 待实现"""
        response = client.get("/api/auth/me")
        assert response.status_code == 501
        assert response.json()["detail"] == "获取当前用户待实现"

    def test_get_current_user_no_auth(self, client):
        """测试未授权访问"""
        response = client.get("/api/auth/me")
        assert response.status_code == 501


class TestProfileUpdate:
    """资料更新测试"""

    def test_update_profile_returns_501(self, client):
        """测试更新资料返回501 - 待实现"""
        response = client.put(
            "/api/auth/profile",
            json={
                "openid": "test_openid",
                "nickname": "新昵称"
            }
        )
        assert response.status_code == 501
        assert response.json()["detail"] == "更新用户信息待实现"


class TestStudentProfile:
    """学生画像测试"""

    def test_create_student_profile_returns_501(self, client):
        """测试创建学生画像返回501 - 待实现"""
        response = client.post(
            "/api/auth/profile/student",
            json={
                "major": "计算机科学与技术",
                "grade": 2
            }
        )
        assert response.status_code == 501
        assert response.json()["detail"] == "创建学生画像待实现"

    def test_get_student_profile_returns_501(self, client):
        """测试获取学生画像返回501 - 待实现"""
        response = client.get("/api/auth/profile/student")
        assert response.status_code == 501
        assert response.json()["detail"] == "获取学生画像待实现"


class TestAdminAuth:
    """管理员认证测试"""

    def test_admin_login_returns_501(self, client):
        """测试管理员登录返回501 - 待实现"""
        response = client.post(
            "/api/auth/admin/login",
            json={
                "username": "admin",
                "password": "admin123"
            }
        )
        assert response.status_code == 501
        assert response.json()["detail"] == "管理员登录待实现"

    def test_get_current_admin_returns_501(self, client):
        """测试获取当前管理员返回501 - 待实现"""
        response = client.get("/api/auth/admin/me")
        assert response.status_code == 501
        assert response.json()["detail"] == "获取管理员信息待实现"


class TestAuthSchemas:
    """认证Schema测试"""

    def test_user_login_schema(self):
        """测试UserLogin模型"""
        from app.schemas.auth import UserLogin
        login = UserLogin(code="test_code")
        assert login.code == "test_code"

    def test_user_create_schema(self):
        """测试UserCreate模型"""
        from app.schemas.auth import UserCreate
        user = UserCreate(openid="test_openid")
        assert user.openid == "test_openid"
        assert user.nickname is None

    def test_token_response_schema(self):
        """测试TokenResponse模型"""
        from app.schemas.auth import TokenResponse
        token = TokenResponse(
            access_token="test_token",
            expires_in=10080
        )
        assert token.access_token == "test_token"
        assert token.token_type == "bearer"
        assert token.expires_in == 10080


class TestUnauthorizedAccess:
    """未授权访问测试"""

    def test_protected_endpoint_without_token(self, client):
        """测试无Token访问受保护端点"""
        # /api/auth/me 是受保护端点，当前返回501因为功能未实现
        response = client.get("/api/auth/me")
        assert response.status_code == 501

    def test_protected_endpoint_with_invalid_token(self, client):
        """测试无效Token访问受保护端点"""
        headers = {"Authorization": "Bearer invalid_token"}
        response = client.get("/api/auth/me", headers=headers)
        # 当前返回501因为功能未实现，未来实现后应返回401
        assert response.status_code == 501


if __name__ == "__main__":
    pytest.main([__file__, "-v"])